Add support for various hash algorithms for specifying checksums (#300)
This PR allows the user to either specify (or not specify) the hash
algorithm as a prefix to the checksum within conf/checksums. Further
this also expands on the possible algorithms which can be used to
compute and specify checksums. Supported algorithms are sha512, sha384
sha256, sha224, sha1, md5. If the algorithm is not specified in the
checksums file, it is inferred based on the length of the hash.