Update influxdb checksum (#412)
The influxdb checksum appears to have changed again (see #381)
due to re-signing the RPM. The contents of the RPM were verified
to not have changed. Both the RPM scriptlets/metadata and the CPIO
payload are identical, but the GPG signature has a newer timestamp.

To make it clear that the checksum provided is for the influxdb package
found under https://repos.influxdata.com/centos/7/x86_64/stable/,
specify the entire URL in the defaults file rather then specify only the
name and include the rest of the URL in the get_url task.