Add optional Application Insights on azure
This PR allows for the Azure Application Insights Java agent to be
configured to run with the manager and tablet servers. When specified,
this will create an Application Insights object along with the cluster,
and metrics will be sent there. The default configuration sends
micrometer metrics, but the user could configure differently (e.g., to
send certain JMX metrics).

* Create workbook if either oms integration or app insights are used.
* Create application insights resource as part of azure cluster setup
* Configure Accumulo to run the application insights Java agent
* Clean up parameter to azureDeployLinuxCounters.json template to remove
  duplication of the creation of the log analytics workspace name.
* Save the log analytics workspace resource ID in muchos.props along
  with the workspace ID and key, since this value is needed to set up
  the application insights resource.