Use specified image for Azure Proxy VM
Rather than having a hard-coded value for the SKU/image to use for the
Azure proxy VM, default to using the value specified for the rest of the
cluster. The value can be overridden specifically for the proxy by
setting azure_proxy_image_reference in the azure section of the config.