Use cloud-init for Azure deployments
Define and use a cloud-init config which is then passed to the Azure
instance creation modules as custom data. Currently the only
per-instance configuration specified is to upgrade the existing CentOS
packages on the VM. This provides a Azure-specific solution to a
Azure-specific requirement (as there is not a more recent version
of CentOS 7.9 for Azure). Without this, currently Muchos cannot
be used to setup a cluster on Azure as the CA certificates are outdated
and prevent Ansible from downloading tarballs on the proxy host.