Use cloud-init for Azure deployments (#420)
Define and use a cloud-init config which is then passed to the Azure
instance creation modules as custom data. Currently the only
per-instance configuration specified is to upgrade the existing CentOS
packages on the VM. This also provides an interim Azure-specific 
solution to an Azure-specific requirement (as there is not a more recent 
version of CentOS 7.x for Azure). Without this, currently Muchos cannot
be used to setup a cluster on Azure as the CA certificates are outdated
and prevent Ansible from downloading tarballs on the proxy host.

Future work will be done to evaluate and potentially use a more modern
default image for Azure, as well as generalize using cloud-init  across 
both EC2 and Azure cluster types in Muchos.