Avoid using Ansible module to download javax.activation-api (#432)
In #294 we had added support for Hadoop (<= 3.2.3) on Java 11 by
downloading and copying the javax.activation-api artifact using
Ansible's `maven_artifact` module. This module in turn needs `lxml`
which needed to be installed separately. The installation of `lxml`
ended up with variations (CentOS 7/8, Fedora) etc.

So this PR simplifies this by directly calling the Maven binary to
download the artifact on the proxy for later copying to the individual
Hadoop nodes.