Update Ansible version and consequent affected files (#433)
The erstwhile `ansible-core` module was flagged as having a
[vulnerability](https://github.com/apache/fluo-muchos/security/dependabot/1)
which is fixed in ansible-core 2.15. After upgrading to that version of
Ansible, unfortunately a bunch of other new ansible-lint warnings needed
to be addressed. Hence this PR looks big, but the changes are largely
stylistic. I will follow this PR up with another related PR specific to
the Azure related ansible modules which also needed update, but this PR
should resolve the Dependabot alert.